requirements.txt
streamlit>=1.27.0
openai>=1.3.0
anthropic>=0.3.0
google-generativeai>=0.2.0
mistralai>=0.0.7
transformers>=4.30.0
torch>=2.0.0
bert-score>=0.3.13
nltk>=3.8.1
spacy>=3.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
rouge-score>=0.1.2
tenacity>=8.2.0 